requirements.txt
PyExecJS
requests
loguru
python-dotenv
retry
openpyxl
pandas
openai